Refactor to move common vp8 functions into their own file to make it nicer to use
